assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    # 4110 only:
    # "cython>=3",
]
build-backend = "setuptools.build_meta"

[project]
version = "1.0.0"
requires-python = ">=3.8"
license = {text = "MIT License"}
name = "in3110_instapy"
description = "color scale"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "line-profiler",
    "pillow"
]


[project.scripts]
instapy = "in3110_instapy.cli:main"
